modules/google-earth-engine/requirements.txt
git+git://github.com/openforis/earthengine-api.git@v0.1.270#egg=earthengine-api&subdirectory=python
git+git://github.com/Servir-Mekong/landcoverPackage#egg=servir-mekong-land-cover
Flask>=0.12.2
git+git://github.com/cdanielw/promise
python-dateutil>=2.7.2
requests>=2.20.0
gunicorn==19.7.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
google-api-python-client==1.7.9
git+git://github.com/cdanielw/RxPY.git@3.0.1#egg=rx
